requirements.txt
ollama==0.4.7
openai==1.61.0
python-dotenv==1.0.1